Project layout refactor, gitignore update.
