+++ .gitignore
@@ -1,3 +1,6 @@
 
 /build*
 /external*
+/image*
+*.img
+*_serial.txt
