+++ kernel/include/atomic.h
