+++ kernel/serial.c
@@ -1,5 +1,5 @@
 
-#include <kernel/io.h>
+#include "io.h"
 #include "serial.h"
 
 static void init_port(short port)
